requirements.txt
numpy>=1.20.0,<2.0.0
Pillow>=9.0.0
matplotlib>=3.5.0
